requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.5
django-allauth==0.51.0
django-gitlab-auth==0.1.1
django-js-asset==2.0.0
django-mptt==0.13.4
idna==3.3
oauthlib==3.2.0
Pillow==9.1.1
pycparser==2.21
PyJWT==2.4.0
python-social-auth==0.3.6
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
